Cache the bug detail page
We can do this with no timeout because we have the bug's modified_at
timestamp, which changes whenever any action is done on the bug.

If we ever implement editing of titles, this will be problematic,
because bug mentions include the other bug's title. If the other bug's
title gets updated, the cached rendered markdown won't be updated.